+++ rlbot.cfg
@@ -19,16 +19,16 @@ enable_state_setting = True
 [Participant Configuration]
 # Put the name of your bot config file here.  Only num_participants config files will be read!
 # Everything needs a config, even players and default bots.  We still set loadouts and names from config!
-participant_config_0 = src/bot.cfg
-participant_config_1 = src/bot.cfg
-participant_config_2 = src/bot.cfg
-participant_config_3 = src/bot.cfg
-participant_config_4 = src/bot.cfg
-participant_config_5 = src/bot.cfg
-participant_config_6 = src/bot.cfg
-participant_config_7 = src/bot.cfg
-participant_config_8 = src/bot.cfg
-participant_config_9 = src/bot.cfg
+participant_config_0 = src/scenarios/bot.cfg
+participant_config_1 = src/scenarios/bot.cfg
+participant_config_2 = src/scenarios/bot.cfg
+participant_config_3 = src/scenarios/bot.cfg
+participant_config_4 = src/scenarios/bot.cfg
+participant_config_5 = src/scenarios/bot.cfg
+participant_config_6 = src/scenarios/bot.cfg
+participant_config_7 = src/scenarios/bot.cfg
+participant_config_8 = src/scenarios/bot.cfg
+participant_config_9 = src/scenarios/bot.cfg
 
 # team 0 shoots on positive goal, team 1 shoots on negative goal
 participant_team_0 = 0
