+++ src/scenarios/goalie/goalie_bot.cfg
@@ -1,15 +1,19 @@
 [Locations]
 # Path to loadout config. Can use relative path from here.
-looks_config = ./appearance.cfg
+looks_config = ../appearance.cfg
 
 # Path to python file. Can use relative path from here.
-python_file = ./bot.py
+python_file = ../../bots/model_bot.py
 
 # Name of the bot in-game
 name = PGBot
 
 # The maximum number of ticks per second that your bot wishes to receive.
 maximum_tick_rate_preference = 120
+
+[Bot Parameters]
+# Path for model
+model_path = ../../models/GoalKeeping.onnx
 
 [Details]
 # These values are optional but useful metadata for helper programs
