+++ src/scenarios/goalie/goalie_match.cfg
