+++ src/scenarios/goalie/goalie_training.py
@@ -7,9 +7,9 @@ from rlbottraining.rng import SeededRandomNumberGenerator
 from rlbottraining.training_exercise import Playlist
 from rlbot.utils.game_state_util import GameState, BallState, CarState, Physics, Vector3, Rotator
 from rlbottraining.common_graders.compound_grader import CompoundGrader
-from base_graders import PassOnBallGoingAwayFromGoal, PassOnGoalForAllyTeam, PassOnTimeout
+from src.graders.base_graders import PassOnBallGoingAwayFromGoal, PassOnGoalForAllyTeam, PassOnTimeout
 
-from src.util.vec import Vec3
+from src.utils.vec import Vec3
 
 
 class GoalieGrader(CompoundGrader):
