+++ src/utils/drive.py
@@ -2,8 +2,8 @@ import math
 
 from rlbot.utils.structures.game_data_struct import PlayerInfo
 
-from util.orientation import Orientation, relative_location
-from util.vec import Vec3
+from .orientation import Orientation, relative_location
+from .vec import Vec3
 
 
 def limit_to_safe_range(value: float) -> float:
