+++ src/utils/orientation.py
@@ -1,6 +1,6 @@
 import math
 
-from util.vec import Vec3
+from .vec import Vec3
 
 
 # This is a helper class for calculating directions relative to your car. You can extend it or delete if you want.
