--- training/README.md
@@ -1,13 +1,0 @@
-# Training
-
-This folder contains exercise scenarios and training runners.
-
-## Exercises
-
-Following exercise scenarios are currently implemented:
-- Goalie Exercise (ball rolling to goal)
-
-## Usage
-
-Create a virtual environment / conda environment and install `pip`. Run `pip install -r requirements.txt` or `pip install rlbottraining`. After that
-a training runner can be executed by running e.g. `rlbottraining run_module training\goalie_runner.py` or `python traning\goalie_runner.py` from the root directory.
